+++ tools/tools.go
@@ -5,6 +5,8 @@
 package tools
 
 import (
+	_ "github.com/bufbuild/buf/cmd/buf"
+	_ "github.com/golang/protobuf/protoc-gen-go"
 	_ "github.com/golangci/golangci-lint/cmd/golangci-lint"
 	_ "github.com/hashicorp/go-hclog/hclogvet"
 	_ "gotest.tools/gotestsum"